Adding a conditional to check method beforehand

+++ api/v1/views/states.py
@@ -17,8 +17,9 @@ def allObjs():
     """
     Get all objs from State
     """
-    result = [objs.to_dict() for objs in storage.all('State').values()]
-    return jsonify(result)
+    if request.method == 'GET':
+        result = [objs.to_dict() for objs in storage.all('State').values()]
+        return jsonify(result)
 
 
 @app_views.route('/states/<state_id>', methods=['GET'], strict_slashes=False)
@@ -26,11 +27,12 @@ def getId(state_id):
     """
     Get the id of the states
     """
-    st = storage.get('State', state_id)
+    if request.method == 'GET':
+        st = storage.get('State', state_id)
 
-    if st is None:
-        abort(404)
-    return jsonify(state.to_dict())
+        if st is None:
+            abort(404)
+        return jsonify(state.to_dict())
 
 
 @app_views.route('/states/<state_id>', methods=['DELETE'],
@@ -39,13 +41,14 @@ def delState(state_id):
     """
     delete a state obj given by state_id
     """
-    st = storage.get('State', state_id)
+    if request.method == 'DELETE':
+        st = storage.get('State', state_id)
 
-    if st is None:
-        abort(404)
-    st.delete()
-    storage.save()
-    return jsonify({}), 200
+        if st is None:
+            abort(404)
+        st.delete()
+        storage.save()
+        return jsonify({}), 200
 
 
 @app_views.route('/states/', methods=['POST'], strict_slashes=False)
@@ -53,17 +56,17 @@ def createState():
     """
     Create a new State object
     """
-    if not request.get_json():
-        return jsonify({
-            'error': 'Not a JSON'
-        }), 400
+    if request.method == 'PUT':
+        if not request.get_json():
+            return jsonify({
+                'error': 'Not a JSON'
+            }), 400
 
-    elif 'name' not in request.get_json():
-        return jsonify({
-            'error:' 'Missing name',
-            400
-        })
-    else:
+        elif 'name' not in request.get_json():
+            return jsonify({
+                'error:' 'Missing name',
+                400
+            })
         objData = request.get_json()
         obj = State(**objData)
         obj.save()
@@ -75,14 +78,17 @@ def updateState(states_id):
     """
     Update an existing state object
     """
-    if not request.get_json():
-        return jsonify({
-            'error': 'Not a JSON'
-        }), 400
-    obj = storage.get('State', states_id)
-    if obj is None:
-        abort(404)
-    objData = request.get_json()
-    objName = objData['name']
-    obj.save()
-    return jsonify(obj.to_dict()), 200
+    if request.method == 'PUT':
+        if not request.get_json():
+            return jsonify({
+                'error': 'Not a JSON'
+            }), 400
+
+        obj = storage.get('State', states_id)
+
+        if obj is None:
+            abort(404)
+        objData = request.get_json()
+        objName = objData['name']
+        obj.save()
+        return jsonify(obj.to_dict()), 200
